Implement especifiers ' ' and '+' after %
A flags strcuture array was created to storage which flags are being activated
that structure is handled as a parameter to every place_x function

a function called putInBuffer was created to save up code. This function is used
in every place_x function. Althought it has not been implemented in all function
yet